+++ packages/ad_sdk/test/native_ad_widget_test.dart
@@ -15,6 +15,9 @@ class _NativeCountingAdapter implements AdProviderAdapter {
   final AdSlot interstitialSlot = AdSlot(type: AdSlotType.interstitial);
   @override
   final AdSlot rewardedSlot = AdSlot(type: AdSlotType.rewarded);
+  @override
+  final AdSlot rewardedInterstitialSlot =
+      AdSlot(type: AdSlotType.rewardedInterstitial);
   final AdSlot _bannerSlot = AdSlot(type: AdSlotType.banner);
   @override
   AdSlot bannerSlot(Object key) => _bannerSlot;
@@ -85,6 +88,8 @@ class _NativeCountingAdapter implements AdProviderAdapter {
   Future<void> loadInterstitial() async {}
   @override
   Future<void> loadRewarded() async {}
+  @override
+  Future<void> loadRewardedInterstitial() async {}
   @override
   Future<void> loadAppOpen({void Function(bool)? onAdLoaded}) async {}
   @override